refactor(logs): unify PostHogLogLevel/Severity, align defaults, add flushAt
- Collapse PostHogLogLevel (replay 3-level) and PostHogLogSeverity (logs
  6-level) into one PostHogLogLevel with 6 cases ordered by severity
  (trace < debug < info < warn < error < fatal). Drop the iOS-only
  guard so the type is cross-platform. Replay case names unchanged;
  only consumers reading raw values directly are affected.
- config.logs.maxBufferSize default now reads from
  PostHogConfig.Defaults.maxQueueSize (1000 / 100 on tvOS) instead of
  the hardcoded 100.
- New config.logs.flushAt knob separates the flush threshold from
  maxBatchSize. Default matches config.flushAt.
