+++ PostHog/Logs/PostHogLogRecord.swift
@@ -14,7 +14,7 @@ import Foundation
     @objc public var body: String
 
     /// Severity level. Defaults to `.info`.
-    @objc public var level: PostHogLogSeverity
+    @objc public var level: PostHogLogLevel
 
     /// Optional attributes attached to the record. Values must be JSON-serializable;
     /// `nil` values are filtered out before sending.
@@ -76,7 +76,7 @@ import Foundation
 
     public init(
         body: String,
-        level: PostHogLogSeverity = .info,
+        level: PostHogLogLevel = .info,
         attributes: [String: Any] = [:],
         traceId: String? = nil,
         spanId: String? = nil,
@@ -148,7 +148,7 @@ import Foundation
     static func fromStorageJSON(_ json: [String: Any]) -> PostHogLogRecord? {
         guard let body = json["body"] as? String else { return nil }
         let levelName = (json["level"] as? String) ?? "info"
-        let level = PostHogLogSeverity.from(name: levelName) ?? .info
+        let level = PostHogLogLevel.from(name: levelName) ?? .info
         let attributes = (json["attributes"] as? [String: Any]) ?? [:]
         let timeUnixNano = (json["timeUnixNano"] as? String) ?? nanosNow()
         let observedTimeUnixNano = (json["observedTimeUnixNano"] as? String) ?? timeUnixNano
